+++ wfc3tools/sub2full.py
@@ -1,12 +1,87 @@
+"""
+sub2full:
+
+    Given an image specified by the user which contains a subarray readout,
+    return the location of the corner of the subarray in a full frame reference
+    image (including the full physical extent of the chip), in 1-indexed
+    pixels. If the user supplies an X and Y coordinate, then the translated
+    location of that point will be returned.
+
+Usage:
+
+    >>> from wfc3tools import sub2full
+    >>> sub2full('ibbso1fdq_flt.fits')
+    >>> [(3584.0, 1539)]
+
+    Specify a list of images:
+
+        >>> im = ['ic5p02e0q_spt.fits',
+                'ic5p02e1q_spt.fits',
+                'ic5p02e2q_spt.fits',
+                'ic5p02e3q_spt.fits',
+                'ic5p02e4q_spt.fits']
+
+        >>> sub2full(im)
+        >>> [(1062.0, 1363),
+            (1062.0, 1363),
+            (1410.0, 1243),
+            (1410.0, 1243),
+            (1402.0, 1539)]
+
+    Return the full extent of the subarray:
+
+        >>> sub2full('ibbso1fdq_flt.fits',fullExtent=True)
+        >>> [(3584.0, 4096, 1539, 2050)]
+
+"""
+
 # STDLIB
 from astropy.io import fits
 import os
 from stsci.tools import parseinput
 
-__taskname__ = "sub2full"
-
 
 def sub2full(filename, x=None, y=None, fullExtent=False):
+    """
+    Given an image specified by the user which contains a subarray readout,
+    return the location of the corner of the subarray in a full frame reference
+    image (including the full physical extent of the chip), in 1-indexed
+    pixels.
+
+    Parameters
+    ----------
+    filename : str or list
+        Input image name or list of image names. The rootname will be used to
+        find the _SPT files in the same directory, the SPT file has all the
+        necessary information for the transform.
+
+    x : int, optional, default=None
+        Specify an x coordinate in the subarray to translate. If an x and y are
+        specified, the fullExtent option is turned off and only the translated
+        x,y coords are returned.
+
+    y : int, optional, default=None
+        Specify a y coordinate in the subarray to translate. If an x and y are
+        specified, the fullExtent option is turned off and only the translated
+        x,y coords are returned.
+
+    fullExtent : bool, optional, default=False
+        If True, the returned values will include the full extent of the
+        subarray in the reference image, for example: (x0,x1,y0,y1).
+
+    Returns
+    -------
+    coords : list
+        A list of tuples which specify the translated coordinates, either
+        (x0,y0) for each image or the full extent sections.
+
+    Examples
+    --------
+    >>> from wfc3tools import sub2full
+    >>> filename = 'ibbso1fdq_flt.fits'
+    >>> coords = sub2full(filename, x=None, y=None, fullExtent=False)
+
+    """
 
     infiles, dummy_out = parseinput.parseinput(filename)
     if len(infiles) < 1:
@@ -15,7 +90,8 @@ def sub2full(filename, x=None, y=None, fullExtent=False):
     coords = list()
 
     for f in infiles:
-        spt = f[0:9] + '_spt.fits'
+        spt = os.path.join(os.path.dirname(f), os.path.basename(f)[0:9] +
+                           '_spt.fits')
         uvis_x_size = 2051
         serial_over = 25.0
         ir_overscan = 5.0